myAlexa/requirements.txt
SpeechRecognition
pyttsx3
PyAudio
pywhatkit
wikipedia
python-dotenv
langchain-openai
openai
pandas
PyQt5